\name{SGPdata-package}
\alias{SGPdata-package}
\docType{package}
\title{Exemplar data sets for SGP analyses}
\description{
	The package SGPdata contains three exemplar data sets utilized by the SGP Package for testing and examples.
}

\details{
\tabular{ll}{
Package: \tab SGPdata\cr
Type: \tab Package\cr
Version: \tab 9.0-0.0\cr
Date: \tab 2014-4-1\cr
License: \tab GPL-2\cr
LazyLoad: \tab Yes\cr
LazyData: \tab Yes\cr
}
}

\keyword{package}
