\name{SGPdata-package}
\alias{SGPdata-package}
\alias{SGPdata}
\docType{package}
\title{Exemplar Data Sets for Student Growth Percentile (SGP) Analyses}
\description{
	The 'SGPdata' package contains five exemplar data sets utilized by the 'SGP' package for testing and examples.
}

\details{
\tabular{ll}{
Package: \tab SGPdata\cr
Type: \tab Package\cr
Version: \tab 28.9-3.0\cr
Date: \tab 2025-5-6\cr
License: \tab GPL-3\cr
LazyLoad: \tab Yes\cr
LazyData: \tab Yes\cr
}
}

\keyword{package}
