\name{SGPdata-package}
\alias{SGPdata-package}
\docType{package}
\title{Exemplar data sets for SGP analyses}
\description{
	The package SGPdata contains two exemplar data sets utilized by the SGP Package for testing and examples.
}

\details{
\tabular{ll}{
Package: \tab SGPdata\cr
Type: \tab Package\cr
Version: \tab 1.0-0.0\cr
Date: \tab 2012-1-9\cr
License: \tab CC BY-SA 3.0 US + file LICENSE\cr
LazyLoad: \tab Yes\cr
LazyData: \tab Yes\cr
}
}

\keyword{package}
