\name{SGPdata-package}
\alias{SGPdata-package}
\docType{package}
\title{Exemplar Data Sets for SGP Analyses}
\description{
	The package SGPdata contains four exemplar data sets utilized by the SGP Package for testing and examples.
}

\details{
\tabular{ll}{
Package: \tab SGPdata\cr
Type: \tab Package\cr
Version: \tab 12.0-0.0\cr
Date: \tab 2015-10-20\cr
License: \tab GPL-3\cr
LazyLoad: \tab Yes\cr
LazyData: \tab Yes\cr
}
}

\keyword{package}
